% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_by_decade-saved.R, R/mhw_by_decade.R
\name{avg_duration_by_decade_sql}
\alias{avg_duration_by_decade_sql}
\title{calculate av duration grouped by decade}
\usage{
avg_duration_by_decade_sql(mhw_table, use_end_date = FALSE)

avg_duration_by_decade_sql(mhw_table, use_end_date = FALSE)
}
\arguments{
\item{mhw_table}{character name of scenario table to use}

\item{use_end_date}{boolean TRUE: use waves ending in the decade; FALSE: use waves starting in the deacde}
}
\value{
character sql code to pull data

character sql code to run
}
\description{
sql code to calc avg duration grouped by decade. Decade start/end dates come
from a table 'decades' that must exist in the database
}
