% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_db.R
\name{check_mhw_connection}
\alias{check_mhw_connection}
\title{check if connection works}
\usage{
check_mhw_connection(con)
}
\description{
test db connection by attempting to run SQL that should work, specific to this project
}
