% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_partitioning.R
\name{create_partion_table}
\alias{create_partion_table}
\title{create table of start/end dates for year gaps}
\usage{
create_partion_table(
  partition_size_years = 3,
  start_year = 2040,
  end_year = 2069
)
}
\description{
create table of start/end dates for year gaps
}
