% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_partitioning.R
\name{create_partition_df}
\alias{create_partition_df}
\title{create table of start/end dates for year gaps}
\usage{
create_partition_df(
  partition_size_years = 3,
  start_year = 2040,
  end_year = 2069
)
}
\arguments{
\item{partition_size_years}{integer width of partitions in whole years, not more than 10 usually}

\item{start_year}{integer year to start partitioning (low end of range), default 2040}

\item{end_year}{integer year to end portioning}
}
\value{
data frame of with years, start_date, end_date
}
\description{
create table of start/end dates for year gaps
}
