% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_db.R
\docType{data}
\name{dbfile}
\alias{dbfile}
\title{path to database}
\format{
An object of class \code{character} of length 1.
}
\usage{
dbfile
}
\description{
path to the database that will be used for defaults, loaded from environment
create a file `.Renviron` in main directory to use that, or in your own R project
}
\keyword{datasets}
