% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning.R
\docType{data}
\name{decades}
\alias{decades}
\title{main decades to use for analysis}
\format{
An object of class \code{character} of length 3.
}
\usage{
decades
}
\description{
main decades to use for analysis
}
\keyword{datasets}
