% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_by_decade.R
\name{duration_by_decade_sql}
\alias{duration_by_decade_sql}
\title{sql for parititioning wave durations by decade}
\usage{
duration_by_decade_sql(mhw_table, use_end_date = FALSE)
}
\arguments{
\item{mhw_table}{character name of scenario table to use}

\item{use_end_date}{boolean TRUE: use waves ending in the decade; FALSE: use waves starting in the deacde}
}
\value{
character sql code to run
}
\description{
collect all durations (not summarized) by decade, useful for histograms. Decade start/end dates come
from a table 'decades' that must exist in the database
}
