% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_decade_plot.R
\name{generate_duration_plots}
\alias{generate_duration_plots}
\title{plot a series of MHW tables as maps and histograms}
\usage{
generate_duration_plots(
  mhw_tables,
  conn,
  threshold = 365/2,
  pdf_file_name = NULL
)
}
\arguments{
\item{mhw_tables}{vector of character names of tables to plot}

\item{conn}{database connection to use with mhw functions}

\item{threshold}{numeric common value to use to constrain scale of all plots, see plot_rasters_squish_outliers, default 365/2}

\item{pdf_file_name}{optional character if set, will be used as a path to a pdf file}
}
\value{
plots or path to plots
}
\description{
plot a series of MHW tables as maps and histograms
}
