% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_db.R
\name{get_dbfile}
\alias{get_dbfile}
\title{path to database}
\usage{
get_dbfile()
}
\description{
path to the database that will be used for defaults, loaded from environment
create a file `.Renviron` in main directory to use that, or in your own R project
}
