% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_by_decade.R
\name{metric_by_decade_sql}
\alias{metric_by_decade_sql}
\title{SQL template for all MHW metrics}
\usage{
metric_by_decade_sql(
  mhw_table,
  mhw_metric = "int_mean",
  sql_function = "avg",
  start_year = 2040,
  end_year = 2069,
  ensemble_list_string = NA
)
}
\arguments{
\item{mhw_table}{name of the table to use}

\item{mhw_metric}{character name of column in table, mhw_dur, int_mean, etc}

\item{start_year}{integer year to start, inclusive (dates will include this start year)}

\item{end_year}{integer year to end, inclusive (dates will include up to 12/31 of the end year)}

\item{ensemble_list_string}{character string that is a list of ensembles to include in format "006,007,008" 
ensembles have leaving zeros, are 3 digits and this must be a separated list with leading zeros}

\item{sqlfun}{character aggregate function for duckdb, see https://duckdb.org/docs/sql/functions/aggregates}
}
\value{
character SQL code to run on the MHW database
}
\description{
this is a convenience function to construct sql to group by decades for one of the metrics
calculating the decade and grouping is tricky so that wraps that.
}
