% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_by_decade.R
\name{metrics_by_decade_raster}
\alias{metrics_by_decade_raster}
\title{rasters of global heatwave duration per decade}
\usage{
metrics_by_decade_raster(
  mhwdb_conn,
  mhw_table,
  mhw_metric = "int_mean",
  sql_function = "avg",
  decades = c("2040", "2050", "2060"),
  ensemble_list_string = NA,
  crs = "EPSG:4087"
)
}
\arguments{
\item{mhwdb_conn}{database connection}

\item{mhw_table}{character string name of the table to query, required (e.g. arise10_metrics)}

\item{ensemble_list_string}{character string that is a list of ensembles to include in format "006,007,008" 
ensembles have leaving zeros, are 3 digits and this must be a separated list with leading zeros}

\item{crs}{character coordinate reference system, EPSG:4087, alternatives are EPSG:4326.}
}
\value{
terra raster stack of layers for each decade
}
\description{
a raster with layers by decade
}
