% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_db.R
\name{mhw_connect}
\alias{mhw_connect}
\title{get connection to mhw database}
\usage{
mhw_connect(duckdbfilepath, required_table_name = "ensembles")
}
\arguments{
\item{duckdbfilepath}{posix path to .duckdb file}

\item{required_table_name}{name of table that must be in db to check for}
}
\value{
duckdb DBI database connection object for use in dbGetQuery, etc or NA if not found
}
\description{
get connection, but also ensure it has at least one of the required tables
}
