% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_partitioning.R
\name{partition_mhw_events_truncated}
\alias{partition_mhw_events_truncated}
\title{table of truncated mhw metrics by partitions}
\usage{
partition_mhw_events_truncated(
  conn,
  mhw_table_name,
  partition_size_years = 3,
  start_year = 2040,
  end_year = 2069
)
}
\arguments{
\item{conn}{dbi database connection, see mhw_connect()}

\item{mhw_table_name}{name of main metrics table to use}

\item{partition_size_years}{width of partions in years}

\item{start_year}{year to start partitions default 2040}

\item{end_year}{year to end partitions  default 2069}
}
\value{
dplyr table with fields from mhw table, partition table (partition_name, partition_start_date, partition_end_date ), 
         mhw_portion = which part of wave, mhw_portion_start =  start date of truncated mhw, 
         mhw_portion_end = end date of truncate wave, partition_dur = duration in days of truncated wave
}
\description{
given set of partions with dates, create table by matching waves to partions and truncated them
}
