% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_partitioning.R
\name{partition_mhw_events_truncated_sql}
\alias{partition_mhw_events_truncated_sql}
\title{sql for partitioning}
\usage{
partition_mhw_events_truncated_sql(
  mhw_table_name,
  partition_size_years = 3,
  start_year = 2040,
  end_year = 2069
)
}
\arguments{
\item{mhw_table_name}{name of main metrics table to use}

\item{partition_size_years}{width of partions in years}

\item{start_year}{year to start partitions default 2040}

\item{end_year}{year to end partitions  default 2069}
}
\value{
character sql to create new table of truncated mwh by partition
}
\description{
sql for partitioning
}
