% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_plot.R
\name{plot_decade_rasters}
\alias{plot_decade_rasters}
\title{plot of world rasters by layer, no outlier management}
\usage{
plot_decade_rasters(mhwdb_conn, mhw_table)
}
\description{
use plot_rasters_squish_outliers instead to plot mhw rasters
}
