% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_summary.R
\name{summary_metrics_raster}
\alias{summary_metrics_raster}
\title{create raster of global marine heatwave summary metric per location}
\usage{
summary_metrics_raster(
  mhwdb_conn,
  mhw_sql,
  mhw_value_name,
  rotate_globe = TRUE,
  crs = "EPSG:4087"
)
}
\arguments{
\item{mhwdb_conn}{database connection}

\item{mhw_sql}{character string of SQL that calculates summary.
SQL must output table with 3 colums (lat, lon, some_column)}

\item{rotate_globe}{= TRUE boolean True mean rotate the global output so prime
meridian is centered, False, no rotation applied, this works if 
data is maped from 0 to 360 instead of -180 to 180}

\item{crs}{= 'EPSG:4087' optional character coordinate reference system}
}
\value{
single terra raster, if
}
\description{
using SQL, for example generated by mhw_metric_summary_sql, creates a single
raster version with lat, long and metric
The raster will have attributes lat, lon and the name of the metric sent
}
