% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_db.R
\name{table_head}
\alias{table_head}
\title{get a few first rows of database table}
\usage{
table_head(conn, tablename, n = 10)
}
\arguments{
\item{conn}{database connection}

\item{tablename}{a table that exists in the database}

\item{n}{number of rows to return, default 10}
}
\value{
dataframe as result of query, or a promise of one for some db
}
\description{
useful for testing/exploration without creating large memory objects
}
