% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_partitioning.R
\name{windowing_overlap}
\alias{windowing_overlap}
\title{create overlapping subsets of a sequence/vector}
\usage{
windowing_overlap(vec, seg.length, overlap = 0)
}
\arguments{
\item{vec}{}

\item{seg.length}{}

\item{overlap}{}
}
\value{
list of subsets
}
\description{
draft/experimental function for creating moving windows of vectors of years
}
