% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_by_decade.R
\name{durations_by_decade_raster}
\alias{durations_by_decade_raster}
\title{rasters of global heatwave duration per decade}
\usage{
durations_by_decade_raster(mhwdb_conn, mhw_table)
}
\arguments{
\item{mhwdb_conn}{database connection}

\item{mhw_table}{character string name of the table to query, required (e.g. arise10_metrics)}
}
\value{
terra raster stack of layers for each decade
}
\description{
a raster with layers by decade
}
