% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_db.R
\docType{data}
\name{mhw_table}
\alias{mhw_table}
\title{get dbplyr table connection}
\format{
An object of class \code{character} of length 1.
}
\usage{
mhw_table
}
\arguments{
\item{conn}{database connection}

\item{table_name}{character name of table}
}
\value{
dbplyr table connection for use in dplyr syntax
}
\description{
get dbplyr table connection
}
\keyword{datasets}
