% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_db.R
\name{mhw_table}
\alias{mhw_table}
\title{get dbplyr table connection}
\usage{
mhw_table(conn, table_name)
}
\arguments{
\item{conn}{database connection}

\item{table_name}{character name of table}
}
\value{
dplyr table connection for use in dplyr syntax or a
}
\description{
get dbplyr table connection
}
