% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhw_partitioning.R
\name{partition_mhw_events}
\alias{partition_mhw_events}
\title{create tabl assigning partition to each event}
\usage{
partition_mhw_events(
  conn,
  mhw_table = "mhw_metrics",
  partition_size_years = 3,
  start_year = 2040,
  end_year = 2069,
  use_end_date = TRUE
)
}
\arguments{
\item{conn}{dbi database connection, see mhw_connect()}

\item{mhw_table}{name of main metrics table to use}

\item{partition_size_years}{width of partions in years}

\item{start_year}{year to start partitions default 2040}

\item{end_year}{year to end partitions  default 2069}

\item{use_end_date}{bolean deafult TRUE}
}
\value{
data frame with partition value assigned to each event
}
\description{
create tabl assigning partition to each event
}
